requirements.txt
ansi==0.2.0
APScheduler==3.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorlog==6.6.0
cryptography==36.0.1
daemonize==2.5.0
deepmerge==1.0.1
dulwich==0.20.30
errbot==6.1.8
Flask==2.0.2
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
pycparser==2.21
Pygments==2.11.2
pygments-markdown-lexer==0.1.0.dev39
pymongo==4.0.1
pyOpenSSL==21.0.0
python-dotenv==0.19.2
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
waitress==2.0.0
WebOb==1.8.7
WebTest==3.0.0
Werkzeug==2.0.2
zipp==3.7.0
